pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cspm"
version = "0.1.0"
description = "Cloud Security Posture Monitor — scans AWS environments for misconfigurations mapped to CIS Benchmarks"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "moto[all]>=5.0.0",
]

[project.scripts]
cspm = "cspm.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
